requirements.txt
numpy
pandas
plotly
statsmodels
streamlit
streamlit_option_menu
seaborn
matplotlib

oauth2client
google-api-python-client
gsheetsdb
google-auth-httplib2
google-auth-oauthlib